% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comethyl.R
\docType{package}
\name{comethyl}
\alias{comethyl}
\title{comethyl: An R package for weighted region comethylation network analysis}
\description{
Comethyl builds upon the \code{WGCNA} package to identify and interpret modules
of comethylated regions from whole-genome bisulfite sequencing data. Regions
are defined from clusters of CpG sites or from genomic annotations, and then
percent methylation values are used to identify comethylation modules.
Interesting modules are identified and explored by comparing with sample traits
and examining functional enrichments. Results are then visualized with
high-quality, editable plots from \code{ggplot2}.
}
\seealso{
GitHub: \url{https://github.com/cemordaunt/comethyl}
}
\author{
Charles Mordaunt
}
