% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_and_Filter_Regions.R
\name{filterRegions}
\alias{filterRegions}
\title{Filter Regions}
\usage{
filterRegions(
  regions,
  covMin = 10,
  methSD = 0.05,
  save = TRUE,
  file = "Filtered_Regions.txt",
  verbose = TRUE
)
}
\arguments{
\item{regions}{A \code{data.frame} output from \code{\link{getRegions()}}
giving the set of regions and statistics for each region.}

\item{covMin}{A \code{numeric(1)} specifying the minimum number of reads at CpGs
in a region in any sample}

\item{methSD}{A \code{numeric(1)} specifying the minimum methylation standard
deviation in a region}

\item{save}{A \code{logical(1)} indicating whether to save the
\code{data.frame}}

\item{file}{A \code{character(1)} giving the file name (.txt) for the saved
\code{data.frame}}

\item{verbose}{A \code{logical(1)} indicating whether messages should be
printed.}
}
\value{
A filtered version of the \code{regions} \code{data.frame}
}
\description{
\code{filterRegions()} filters a region set using specified \code{covMin} and
\code{methSD} thresholds and then saves it as a tab-delimited text file.
}
\details{
The purpose of this function is to use cutoffs for minimum coverage
and methylation standard deviation guided by other functions and obtain a
robust set of variably methylated regions. Computational resources are also a
consideration for network construction, with region sets of 250K or less
generally performing well.
}
\examples{
\dontrun{

# Call Regions
regions <- getRegions(bs, file = "Unfiltered_Regions.txt")
plotRegionStats(regions, maxQuantile = 0.99,
                file = "Unfiltered_Region_Plots.pdf")
plotSDstats(regions, maxQuantile = 0.99,
            file = "Unfiltered_SD_Plots.pdf")

# Examine Region Totals at Different Cutoffs
regionTotals <- getRegionTotals(regions, file = "Region_Totals.txt")
plotRegionTotals(regionTotals, file = "Region_Totals.pdf")

# Filter Regions
regions <- filterRegions(regions, covMin = 10, methSD = 0.05,
                         file = "Filtered_Regions.txt")
plotRegionStats(regions, maxQuantile = 0.99,
                file = "Filtered_Region_Plots.pdf")
plotSDstats(regions, maxQuantile = 0.99,
            file = "Filtered_SD_Plots.pdf")
}

}
\seealso{
\itemize{
\item \code{\link{getRegions()}} to generate the set of regions.
\item \code{\link{plotRegionStats()}}, \code{\link{plotSDstats()}},
\code{\link{getRegionTotals()}}, and
\code{\link{plotRegionTotals()}} for help visualizing region
characteristics and setting cutoffs for filtering.
\item \code{\link{getRegionMeth()}} to get methylation values for
these regions in all samples.
}
}
