% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_and_Filter_CpGs.R
\name{getCpGtotals}
\alias{getCpGtotals}
\title{Get the Total CpGs at Different Coverage Cutoffs}
\usage{
getCpGtotals(
  bs,
  cov = seq(0, 10, 1),
  perSample = seq(0.5, 1, 0.05),
  save = TRUE,
  file = "CpG_Totals.txt",
  verbose = TRUE
)
}
\arguments{
\item{bs}{A \code{\link[bsseq]{BSseq}} object.}

\item{cov}{A \code{numeric} specifying the minimum number of reads
overlapping a CpG for it to be included in the total.}

\item{perSample}{A \code{numeric} specifying the minimum percent of samples
with \code{cov} reads at a CpG for it to be included in the total.}

\item{save}{A \code{logical(1)} indicating whether to save the
\code{data.frame}.}

\item{file}{A \code{character(1)} giving the file name (.txt) for the saved
\code{data.frame}.}

\item{verbose}{A \code{logical(1)} indicating whether messages should be
printed.}
}
\value{
A \code{data.frame} giving the number of CpGs (in millions) and
the percent of total CpGs at all combinations of \code{cov} and
\code{perSample}. The number of samples corresponding to
\code{perSample} is also given.
}
\description{
\code{getCpGtotals()} calculates the total number and percent of CpGs
remaining in a \code{\link[bsseq]{BSseq}} object after filtering at
different \code{cov} and \code{perSample} cutoffs and then saves it
as a tab-separated text file.
}
\details{
The purpose of this function is to help determine cutoffs to maximize the
number of CpGs with sufficient data after filtering. Typically,
the number of CpGs covered in 100\% of samples decreases as the
sample size increases, especially with low-coverage datasets.
}
\examples{
\dontrun{

# Read Bismark CpG Reports
colData <- read.xlsx("sample_info.xlsx", rowNames = TRUE)
bs <- getCpGs(colData, file = "Unfiltered_BSseq.rds")

# Examine CpG Totals at Different Cutoffs
CpGtotals <- getCpGtotals(bs, file = "CpG_Totals.txt")
plotCpGtotals(CpGtotals, file = "CpG_Totals.pdf")

# Filter BSobject
bs <- filterCpGs(bs, cov = 2, perSample = 0.75, file = "Filtered_BSseq.rds")
}

}
\seealso{
\itemize{
\item \code{\link{plotCpGtotals()}} to visualize the CpG totals.
\item \code{\link{filterCpGs()}} to filter the
\code{\link[bsseq]{BSseq}} object.
\item \code{\link{getCpGs()}} to generate the
\code{\link[bsseq]{BSseq}} object from individual Bismark CpG
reports.
}
}
