% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Examine_Region_Methylation.R
\name{getDendro}
\alias{getDendro}
\title{Perform Hierarchical Clustering}
\usage{
getDendro(
  x,
  transpose = FALSE,
  distance = c("euclidean", "pearson", "bicor"),
  maxPOutliers = 0.1,
  verbose = TRUE
)
}
\arguments{
\item{x}{A \code{numeric matrix}.}

\item{transpose}{A \code{logical(1)} specifying whether to transpose the
\code{matrix}.}

\item{distance}{A \code{character(1)} indicating which distance measure to use.
Possible values include \code{euclidean}, \code{pearson}, and
\code{bicor}.}

\item{maxPOutliers}{A \code{numeric(1)} specifying the maximum percentile that
can be considered outliers on each side of the median for the
\code{bicor} statistic.}

\item{verbose}{A \code{logical(1)} indicating whether messages should be
printed.}
}
\value{
An \link[stats:hclust]{stats::hclust} object that describes the clustering tree.
}
\description{
\code{getDendro()} computes the distance between the rows of a matrix and
performs hierarchical clustering. Possible distance measures include
\code{euclidean}, \code{pearson}, and \code{bicor}. The function also
optionally transposes the matrix.
}
\details{
Euclidean distance is calculated by \code{\link[stats:dist]{stats::dist()}}, where
\code{method = "euclidean"}, while Pearson correlation and biweight
midcorrelation (bicor) are computed by \code{\link[WGCNA:cor]{WGCNA::cor()}} and \code{\link[WGCNA:bicor]{WGCNA::bicor()}},
respectively. The \code{cor} and \code{bicor} are then subtracted from 1 to
calculate the dissimilarity. Hierarchical clustering is done by
\code{\link[stats:hclust]{stats::hclust()}}, where \code{method = "average"}.
}
\examples{
\dontrun{

# Assess Sample Similarity
getDendro(methAdj, distance = "euclidean") \%>\%
        plotDendro(file = "Sample_Dendrogram.pdf", expandY = c(0.25,0.08))

# Examine Correlations between Modules
moduleDendro <- getDendro(MEs, distance = "bicor")
plotDendro(moduleDendro, labelSize = 4, nBreaks = 5,
           file = "Module_ME_Dendrogram.pdf")

# Characterize Correlations between Samples
sampleDendro <- getDendro(MEs, transpose = TRUE, distance = "bicor")
plotDendro(sampleDendro, labelSize = 3, nBreaks = 5,
           file = "Sample_ME_Dendrogram.pdf")

# Examine Correlations between Traits
traitDendro <- getCor(MEs, y = colData, corType = "bicor",
                      robustY = FALSE) \%>\%
        getDendro(transpose = TRUE)
plotDendro(traitDendro, labelSize = 3.5, expandY = c(0.65,0.08),
           file = "Trait_Dendrogram.pdf")
}

}
\seealso{
\itemize{
\item \code{\link[=plotDendro]{plotDendro()}} to visualize dendrograms from \code{getDendro()}.
}
}
