% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotate_Modules.R
\name{getGeneList}
\alias{getGeneList}
\title{Extract a Gene List from Annotated Regions}
\usage{
getGeneList(
  regions_annotated,
  module = NULL,
  type = c("symbol", "description", "ensemblID", "entrezID"),
  verbose = TRUE
)
}
\arguments{
\item{regions_annotated}{A \code{data.frame} of regions with gene annotations,
typically produced by \code{\link[=annotateModule]{annotateModule()}}.}

\item{module}{A \code{character} giving the name of one or more modules to
include. If null, all modules will be included.}

\item{type}{A \code{character(1)} specifying the type of gene information to
extract. Possible values include \code{symbol}, \code{description},
\code{ensemblID} and \code{entrezID}.}

\item{verbose}{A \code{logical(1)} indicating whether messages should be
printed.}
}
\value{
A \code{vector} of unique values.
}
\description{
\code{getGeneList()} takes a \code{data.frame} of regions annotated with gene
information and extracts a \code{vector} of unique gene symbols, descriptions,
or IDs.
}
\details{
\code{getGeneList()} is designed to be used in combination with
\code{\link[=annotateModule]{annotateModule()}}. \code{regions} can be filtered for one or more modules of
interest. Values that can be extracted include gene \code{symbol},
\code{description}, \code{ensemblID} and \code{entrezID}.
}
\examples{
\dontrun{

# Get Comethylation Modules
modules <- getModules(methAdj, power = sft$powerEstimate, regions = regions,
                      corType = "pearson", file = "Modules.rds")

# Annotate Modules
regionsAnno <- annotateModule(regions, module = "bisque4",
                              genome = "hg38",
                              file = "Annotated_bisque4_Module_Regions.txt")
geneList_bisque4 <- getGeneList(regionsAnno, module = "bisque4")

# Analyze Functional Enrichment
ontologies <- listOntologies("hg38", version = "4.0.4")
enrich_bisque4 <- enrichModule(regions, module = "bisque4",
                               genome = "hg38",
                               file = "bisque4_Module_Enrichment.txt")
plotEnrichment(enrich_bisque4,
               file = "bisque4_Module_Enrichment_Plot.pdf")
}

}
\seealso{
\itemize{
\item \code{\link[=getModules]{getModules()}} to build a comethylation network and identify
modules of comethylated regions.
\item \code{\link[=annotateModule]{annotateModule()}} to annotate a set of regions with genes and
regulatory context.
\item \code{\link[=listOntologies]{listOntologies()}}, \code{\link[=enrichModule]{enrichModule()}}, and \code{\link[=plotEnrichment]{plotEnrichment()}} to
investigate functional enrichment of module regions with GREAT.
}
}
