% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Explore_Module_Trait_Correlations.R
\name{getMEtraitCor}
\alias{getMEtraitCor}
\title{Calculate Correlation Statistics Between Module Eigennodes and Traits}
\usage{
getMEtraitCor(
  MEs,
  colData,
  corType = c("bicor", "pearson"),
  maxPOutliers = 0.1,
  robustY = FALSE,
  save = TRUE,
  file = "ME_Trait_Correlation_Stats.txt",
  verbose = TRUE
)
}
\arguments{
\item{MEs}{A \code{data.frame} of module eigennode values, where rows are
samples and columns are modules. The row names of \code{MEs} must
include the same set of values as the row names of \code{colData}.}

\item{colData}{A \code{data.frame} whose row names specify samples and whose
columns are sample traits with numeric values.}

\item{corType}{A \code{character(1)} indicating which correlation statistic
to use in the calculation. Potential values include \code{pearson}
and \code{bicor}.}

\item{maxPOutliers}{A \code{numeric(1)} specifying the maximum percentile
that can be considered outliers on each side of the median for the
\code{bicor} statistic.}

\item{robustY}{A \code{logical(1)} indicating whether to use robust
calculation for \code{y} for the \code{bicor} statistic. \code{FALSE}
is recommended if \code{y} is a binary variable.}

\item{save}{A \code{logical(1)} indicating whether to save the
\code{data.frame}.}

\item{file}{A \code{character(1)} giving the file name (.txt) for the saved
\code{data.frame}.}

\item{verbose}{A \code{logical(1)} indicating whether messages should be
printed.}
}
\value{
A \code{data.frame} giving correlation statistics for each
module-trait pair.
}
\description{
\code{getMEtraitCor()} calculates correlation coefficients and p-values
between eigennode values for all modules and all sample traits and saves it
as a .txt file. Correlations are performed using either \code{pearson} or
\code{bicor} methods.
}
\details{
\code{getMEtraitCor()} is designed to be used in combination with
\code{\link[=getModules]{getModules()}}. The correlation calculations are performed by
\code{\link[WGCNA:corAndPvalue]{WGCNA::corAndPvalue()}} and \code{\link[WGCNA:bicorAndPvalue]{WGCNA::bicorAndPvalue()}}. \code{getMEtraitCor()}
can also be used to calculate pairwise correlation coefficients and p-values
between module eigennode values, or between top adjusted PCs and sample
traits (see examples).
}
\examples{
\dontrun{

# Get Comethylation Modules
modules <- getModules(methAdj, power = sft$powerEstimate, regions = regions,
                      corType = "pearson", file = "Modules.rds")

# Test Correlations between Module Eigennodes and Sample Traits
MEs <- modules$MEs
MEtraitCor <- getMEtraitCor(MEs, colData = colData, corType = "bicor",
                            file = "ME_Trait_Correlation_Stats.txt")

# Examine Correlations between Modules
moduleCorStats <- getMEtraitCor(MEs, colData = MEs, corType = "bicor",
                                robustY = TRUE,
                                file = "Module_Correlation_Stats.txt")

# Compare Top PCs to Sample Traits
MEtraitCor <- getMEtraitCor(PCs, colData = colData, corType = "bicor",
                            file = "PC_Trait_Correlation_Stats.txt")

# Examine Correlations between Sample Traits
traitDendro <- getCor(MEs, y = colData, corType = "bicor",
                      robustY = FALSE) \%>\%
        getDendro(transpose = TRUE)
plotDendro(traitDendro, labelSize = 3.5, expandY = c(0.65,0.08),
           file = "Trait_Dendrogram.pdf")

# Visualize Correlations between Module Eigennodes and Sample Traits
moduleDendro <- getDendro(MEs, distance = "bicor")
plotMEtraitCor(MEtraitCor, moduleOrder = moduleDendro$order,
               traitOrder = traitDendro$order,
               file = "ME_Trait_Correlation_Heatmap.pdf")
plotMEtraitCor(MEtraitCor, moduleOrder = moduleDendro$order,
               traitOrder = traitDendro$order, topOnly = TRUE,
               label.type = "p", label.size = 4, label.nudge_y = 0,
               legend.position = c(1.11, 0.795),
               colColorMargins = c(-1,4.75,0.5,10.1),
               file = "Top_ME_Trait_Correlation_Heatmap.pdf", width = 8.5,
               height = 4.25)
}

}
\seealso{
\itemize{
\item \code{\link[=getModules]{getModules()}} to build a comethylation network and identify
modules of comethylated regions.
\item \code{\link[=getCor]{getCor()}} to calculate correlation coefficients.
\item \code{\link[=getDendro]{getDendro()}} and \code{\link[=plotDendro]{plotDendro()}} to generate and visualize
dendrograms.
\item \code{\link[=plotMEtraitCor]{plotMEtraitCor()}} to visualize ME-trait correlations.
}
}
