% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Module_Preservation.R
\name{getModulePreservation}
\alias{getModulePreservation}
\title{Calculate Module Preservation}
\usage{
getModulePreservation(
  meth_disc,
  regions_disc,
  meth_rep,
  regions_rep,
  corType = c("pearson", "bicor"),
  maxPOutliers = 0.1,
  nPermutations = 100,
  save = TRUE,
  file = "Module_Preservation_Stats.txt",
  verbose = TRUE
)
}
\arguments{
\item{meth_disc}{A \code{numeric matrix}, where each row is a sample and each
column is a region. This is typically obtained from
\code{\link[=adjustRegionMeth]{adjustRegionMeth()}} and is related to the discovery (reference) data
set.}

\item{regions_disc}{A \code{data.frame} of regions with module assignments,
which is typically obtained from \code{\link[=getModules]{getModules()}} and is related to the
discovery data set.}

\item{meth_rep}{A \code{numeric matrix}, where each row is a sample and each
column is a region. This is typically obtained from
\code{\link[=adjustRegionMeth]{adjustRegionMeth()}} and is related to the replication (test) data
set.}

\item{regions_rep}{A \code{data.frame} of regions with module assignments,
which is typically obtained from \code{\link[=getModules]{getModules()}} and is related to the
replication data set.}

\item{corType}{A \code{character(1)} indicating which correlation statistic
to use in the adjacency calculation.}

\item{maxPOutliers}{A \code{numeric(1)} specifying the maximum percentile
that can be considered outliers on each side of the median for the
\code{bicor} statistic.}

\item{nPermutations}{A \code{numeric(1)} indicating the number of
permutations to perform in the permutation test.}

\item{save}{A \code{logical(1)} indicating whether to save the
\code{data.frame}.}

\item{file}{A \code{character(1)} giving the file name (.txt) for the saved
\code{data.frame}.}

\item{verbose}{A \code{logical(1)} indicating whether messages should be
printed.}
}
\value{
A \code{data.frame} giving preservation statistics for each module in
the discovery data set.
}
\description{
\code{getModulePreservation()} examines module replication between a
reference and a test data set by estimating various preservation statistics,
which are then saved as a .txt file. Correlations are performed using either
\code{pearson} or \code{bicor} methods.
}
\details{
Identical sets of regions should be assessed and assigned modules within
discovery (reference) and replication (test) data sets, though the replication
regions may be a subset of the discovery regions due to low coverage. It's
also recommended to filter CpGs so identical loci are also assessed within
regions. Network parameters should be as similar as possible, although
modules should be identified independently between the discovery and
replication datasets. Preservation statistics are calculated by
\code{\link[WGCNA:modulePreservation]{WGCNA::modulePreservation()}}, with \code{corFnc} set to either \code{cor} or
\code{bicor}. More information is given in the documentation for
\code{\link[WGCNA:modulePreservation]{WGCNA::modulePreservation()}}.
}
\examples{
\dontrun{
# Calculate Module Preservation
regions_disc <- modules_disc$regions
regions_rep <- modules_rep$regions
preservation <- getModulePreservation(methAdj_disc,
                                      regions_disc = regions_disc,
                                      meth_rep = methAdj_rep,
                                      regions_rep = regions_rep,
                                      corType = "pearson",
                                      file = "Module_Preservation_Stats.txt")

# Visualize Module Preservation
plotModulePreservation(preservation, file = "Module_Preservation_Plots.pdf")
}

}
\seealso{
\itemize{
\item \code{\link[=getModules]{getModules()}} to build a comethylation network and identify
modules of comethylated regions.
\item \code{\link[=plotModulePreservation]{plotModulePreservation()}} to visualize module preservation
statistics.
}
}
