% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Examine_Region_Methylation.R
\name{getRegionMeth}
\alias{getRegionMeth}
\title{Get Region Methylation Data}
\usage{
getRegionMeth(
  regions,
  bs,
  type = c("raw", "smooth"),
  save = TRUE,
  file = "Region_Methylation.rds",
  verbose = TRUE
)
}
\arguments{
\item{regions}{A \code{data.frame} of regions, typically after filtering with
\code{\link{filterRegions()}}. Must have the columns \code{chr},
\code{start}, and \code{end}.}

\item{bs}{A \code{\link[bsseq:`BSseq-class`]{BSseq}} object, typically after
filtering with \code{\link{filterCpGs()}}.}

\item{type}{A \code{character(1)} specifying the type of methylation values
to extract. Accepted values are \code{raw} and \code{smooth}}

\item{save}{A \code{logical(1)} indicating whether to save the \code{matrix}.}

\item{file}{A \code{character(1)} giving the file name (.rds) for the saved
\code{matrix}.}

\item{verbose}{A \code{logical(1)} indicating whether messages should be
printed.}
}
\value{
A \code{numeric matrix}, where each row is a region and each column
is a sample.
}
\description{
\code{getRegionMeth()} extracts methylation values at specified regions for
all samples and then saves it as a .rds file.
}
\details{
Methylation is summarized at the region level, and is estimated as the
methylated reads divided by the total reads. Methylation values are obtained
from a \code{\link[bsseq:`BSseq-class`]{BSseq}} object and can be either raw
or smoothed methylation.
}
\examples{
\dontrun{

# Get Methylation Data
meth <- getRegionMeth(regions, bs = bs, file = "Region_Methylation.rds")

# Adjust Methylation Data for PCs
mod <- model.matrix(~1, data = pData(bs))
methAdj <- adjustRegionMeth(meth, mod = mod,
                            file = "Adjusted_Region_Methylation.rds")

# Assess Sample Similarity
getDendro(methAdj, distance = "euclidean") \%>\%
        plotDendro(file = "Sample_Dendrogram.pdf", expandY = c(0.25,0.08))
}

}
\seealso{
\itemize{
\item \code{\link{adjustRegionMeth()}} to adjust methylation for the
top principal components.
\item \code{\link{getDendro()}} and \code{\link{plotDendro()}} to
generate and visualize dendrograms.
}
}
