% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_and_Filter_Regions.R
\name{getRegionTotals}
\alias{getRegionTotals}
\title{Get Region Totals at Different Cutoffs}
\usage{
getRegionTotals(
  regions,
  covMin = seq(0, 20, 2),
  methSD = seq(0, 0.1, 0.01),
  save = TRUE,
  file = "Region_Totals.txt",
  verbose = TRUE
)
}
\arguments{
\item{regions}{A \code{data.frame} output from \code{\link[=getRegions]{getRegions()}} giving the set
of regions and statistics for each region.}

\item{covMin}{A \code{numeric} specifying the minimum number of reads at CpGs
in a region in any sample for that region to be included in the total.}

\item{methSD}{A \code{numeric} specifying the minimum methylation standard
deviation for that region to be included in the total.}

\item{save}{A \code{logical(1)} indicating whether to save the
\code{data.frame}}

\item{file}{A \code{character(1)} giving the file name (.txt) for the saved
\code{data.frame}}

\item{verbose}{A \code{logical(1)} indicating whether messages should be
printed.}
}
\value{
A \code{data.frame} giving the total number of regions, width, and
number of CpGs at all combinations of \code{covMin} and \code{methSD}.
}
\description{
\code{getRegionTotals()} calculates the total number of regions, as well as
the total width and number of CpGs remaining in a region set after filtering
at different \code{covMin} and \code{methSD} cutoffs and then saves it as a
tab-delimited text file.
}
\details{
The purpose of this function is to help balance cutoffs for minimum coverage
and methylation standard deviation and identify a robust set of variably
methylated regions. It's recommended to input multiple \code{covMin} and
\code{methSD} cutoffs for comparison. Computational resources are also a
consideration for network construction, with region sets of 250K or less
generally performing well.
}
\examples{
\dontrun{

# Call Regions
regions <- getRegions(bs, file = "Unfiltered_Regions.txt")
plotRegionStats(regions, maxQuantile = 0.99,
                file = "Unfiltered_Region_Plots.pdf")
plotSDstats(regions, maxQuantile = 0.99,
            file = "Unfiltered_SD_Plots.pdf")

# Examine Region Totals at Different Cutoffs
regionTotals <- getRegionTotals(regions, file = "Region_Totals.txt")
plotRegionTotals(regionTotals, file = "Region_Totals.pdf")

# Filter Regions
regions <- filterRegions(regions, covMin = 10, methSD = 0.05,
                         file = "Filtered_Regions.txt")
plotRegionStats(regions, maxQuantile = 0.99,
                file = "Filtered_Region_Plots.pdf")
plotSDstats(regions, maxQuantile = 0.99,
            file = "Filtered_SD_Plots.pdf")
}

}
\seealso{
\itemize{
\item \code{\link[=getRegions]{getRegions()}} to generate the set of regions.
\item \code{\link[=plotRegionStats]{plotRegionStats()}}, \code{\link[=plotSDstats]{plotSDstats()}}, and \code{\link[=plotRegionTotals]{plotRegionTotals()}}
for more help visualizing region characteristics and setting
cutoffs for filtering.
\item \code{\link[=filterRegions]{filterRegions()}} for filtering regions by minimum coverage and
methylation standard deviation.
}
}
