% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_Comethylation_Network.R
\name{getSoftPower}
\alias{getSoftPower}
\title{Estimate Soft Power Threshold}
\usage{
getSoftPower(
  meth,
  powerVector = 1:20,
  corType = c("pearson", "bicor"),
  maxPOutliers = 0.1,
  RsquaredCut = 0.8,
  blockSize = 40000,
  gcInterval = blockSize - 1,
  save = TRUE,
  file = "Soft_Power.rds",
  verbose = TRUE
)
}
\arguments{
\item{meth}{A \code{numeric matrix}, where each row is a sample and each
column is a region. This is typically obtained from
\code{\link[=adjustRegionMeth]{adjustRegionMeth()}}.}

\item{powerVector}{A \code{numeric} specifying the soft power thresholds to
examine for scale-free topology.}

\item{corType}{A \code{character(1)} indicating which correlation statistic
to use in the adjacency calculation.}

\item{maxPOutliers}{A \code{numeric(1)} specifying the maximum percentile that
can be considered outliers on each side of the median for the
\code{bicor} statistic.}

\item{RsquaredCut}{A \code{numeric(1)} giving the minimum R-squared value for
scale-free topology. Used to choose the best soft-thresholding power.}

\item{blockSize}{A \code{numeric(1)} specifying the number of regions in each
block for the connectivity calculation. Decrease this if memory is
insufficient.}

\item{gcInterval}{A \code{numeric(1)} indicating the interval for garbage
collection.}

\item{save}{A \code{logical(1)} indicating whether to save the \code{list}.}

\item{file}{A \code{character(1)} giving the file name (.rds) for the saved
\code{list}.}

\item{verbose}{A \code{logical(1)} indicating whether messages should be
printed.}
}
\value{
A \code{list} with two elements: \code{powerEstimate}, which gives the
estimated best soft-thresholding power, and \code{fitIndices}, which
is a \code{data.frame} with statistics on scale-free topology,
including fit and connectivity, along with network density,
centralization, and heterogeneity.
}
\description{
\code{getSoftPower()} analyzes scale-free topology to estimate the best
soft-thresholding power from a vector of powers, calculate fit indices, and
then saves this as a .rds file. Possible correlation statistics include
\code{pearson} and \code{bicor}.
}
\details{
Soft power is estimated by \code{\link[WGCNA:pickSoftThreshold]{WGCNA::pickSoftThreshold()}}, with \code{corFnc}
set to either \code{cor} or \code{bicor}. Calculations are performed for a
signed network in blocks of regions of size \code{blockSize} (default = 40000).
The best soft power threshold is chosen as the lowest power where fit
(R-squared) is greater than \code{RsquaredCut} (default = 0.8). More
information is given in the documentation for \code{\link[WGCNA:pickSoftThreshold]{WGCNA::pickSoftThreshold()}}.
}
\examples{
\dontrun{

# Get Methylation Data
meth <- getRegionMeth(regions, bs = bs, file = "Region_Methylation.rds")

# Adjust Methylation Data for PCs
mod <- model.matrix(~1, data = pData(bs))
methAdj <- adjustRegionMeth(meth, mod = mod,
                            file = "Adjusted_Region_Methylation.rds")

# Select Soft Power Threshold
sft <- getSoftPower(methAdj, corType = "pearson", file = "Soft_Power.rds")
plotSoftPower(sft, file = "Soft_Power_Plots.pdf")

# Get Comethylation Modules
modules <- getModules(methAdj, power = sft$powerEstimate, regions = regions,
                      corType = "pearson", file = "Modules.rds")
}

}
\seealso{
\itemize{
\item \code{\link[=getRegionMeth]{getRegionMeth()}} and \code{\link[=adjustRegionMeth]{adjustRegionMeth()}} to extract methylation
data and then adjust it for the top principal components.
\item \code{\link[=plotSoftPower]{plotSoftPower()}} to visualize fit and connectivity for soft
power estimation.
\item \code{\link[=getModules]{getModules()}} to build a comethylation network and identify
modules of comethylated regions.
}
}
