% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_and_Filter_CpGs.R
\name{plotCpGtotals}
\alias{plotCpGtotals}
\title{Visualize CpG Totals}
\usage{
plotCpGtotals(
  CpGtotals,
  nBreaks = 4,
  legend.position = c(1.08, 0.73),
  save = TRUE,
  file = "CpG_Totals.pdf",
  width = 11,
  height = 4.25,
  verbose = TRUE
)
}
\arguments{
\item{CpGtotals}{A \code{data.frame}, output from \code{\link{getCpGs()}}.}

\item{nBreaks}{A \code{numeric(1)} specifying the number of breaks used for
both axes and the legend.}

\item{legend.position}{A \code{numeric(2)} specifying the position of the
legend, as x-axis, y-axis. May also be a \code{character(1)}
indicating "none", "left", "right", "bottom", or "top".}

\item{save}{A \code{logical(1)} indicating whether to save the plot.}

\item{file}{A \code{character(1)} giving the file name (.pdf) for the saved
plot.}

\item{width}{A \code{numeric(1)} specifying the width in inches of the saved
plot.}

\item{height}{A \code{numeric(1)} specifying the height in inches of the
saved plot.}

\item{verbose}{A \code{logical(1)} indicating whether messages should be
printed.}
}
\value{
A \code{ggplot} object.
}
\description{
\code{plotCpGtotals()} plots the number of CpGs remaining after filtering by
different combinations of \code{cov} and \code{perSample} in a line
plot and then saves it as a pdf.
}
\details{
\code{plotCpGtotals()} is designed to be used in combination with
\code{\link{getCpGtotals()}}. A \code{ggplot} object is produced and
can be further edited outside of this function if desired.
}
\examples{
\dontrun{

# Read Bismark CpG Reports
colData <- read.xlsx("sample_info.xlsx", rowNames = TRUE)
bs <- getCpGs(colData, file = "Unfiltered_BSseq.rds")

# Examine CpG Totals at Different Cutoffs
CpGtotals <- getCpGtotals(bs, file = "CpG_Totals.txt")
plotCpGtotals(CpGtotals, file = "CpG_Totals.pdf")

# Filter BSseq Object
bs <- filterCpGs(bs, cov = 2, perSample = 0.75, file = "Filtered_BSseq.rds")
}

}
\seealso{
\itemize{
\item \code{\link{getCpGs()}} to generate the
\code{\link[bsseq:`BSseq-class`]{BSseq}} object from
individual Bismark CpG reports.
\item \code{\link{getCpGtotals()}} to generate \code{CpGtotals}.
\item \code{\link{filterCpGs()}} to filter the
\code{\link[bsseq:`BSseq-class`]{BSseq}} object.
}
}
