% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Examine_Region_Methylation.R
\name{plotDendro}
\alias{plotDendro}
\title{Plot a Dendrogram}
\usage{
plotDendro(
  dendro,
  label = TRUE,
  labelSize = 2.5,
  expandX = c(0.03, 0.03),
  expandY = c(0.3, 0.08),
  nBreaks = 4,
  save = TRUE,
  file = "Dendrogram.pdf",
  width = 11,
  height = 4.25,
  verbose = TRUE
)
}
\arguments{
\item{dendro}{An \link[stats:hclust]{stats::hclust} object generated by \code{\link[=getDendro]{getDendro()}}.}

\item{label}{A \code{logical(1)} indicating whether to add labels to the
dendrogram.}

\item{labelSize}{A \code{numeric(1)} specifying the size of the label text.}

\item{expandX}{A \code{numeric(2)} giving the x-axis multiplicative range
expansion factors as upper limit, lower limit.}

\item{expandY}{A \code{numeric(2)} giving the y-axis multiplicative range
expansion factors as upper limit, lower limit.}

\item{nBreaks}{A \code{numeric(1)} specifying the number of breaks used for
the y-axis.}

\item{save}{A \code{logical(1)} indicating whether to save the plot.}

\item{file}{A \code{character(1)} giving the file name (.pdf) for the saved
plot.}

\item{width}{A \code{numeric(1)} specifying the width in inches of the saved
plot.}

\item{height}{A \code{numeric(1)} specifying the height in inches of the
saved plot.}

\item{verbose}{A \code{logical(1)} indicating whether messages should be
printed.}
}
\value{
A \code{ggplot} object.
}
\description{
\code{plotDendro()} extracts plotting data from an \link[stats:hclust]{stats::hclust}
object, plots a dendrogram, and then saves it as a .pdf.
}
\details{
\code{plotDendro()} is designed to be used in combination with \code{\link[=getDendro]{getDendro()}}.
A \code{ggplot} object is produced and can be edited outside of this function
if desired.
}
\examples{
\dontrun{

# Assess Sample Similarity
getDendro(methAdj, distance = "euclidean") \%>\%
        plotDendro(file = "Sample_Dendrogram.pdf", expandY = c(0.25,0.08))

# Examine Correlations between Modules
moduleDendro <- getDendro(MEs, distance = "bicor")
plotDendro(moduleDendro, labelSize = 4, nBreaks = 5,
           file = "Module_ME_Dendrogram.pdf")

# Characterize Correlations between Samples
sampleDendro <- getDendro(MEs, transpose = TRUE, distance = "bicor")
plotDendro(sampleDendro, labelSize = 3, nBreaks = 5,
           file = "Sample_ME_Dendrogram.pdf")

# Examine Correlations between Traits
traitDendro <- getCor(MEs, y = colData, corType = "bicor",
                      robustY = FALSE) \%>\%
        getDendro(transpose = TRUE)
plotDendro(traitDendro, labelSize = 3.5, expandY = c(0.65,0.08),
           file = "Trait_Dendrogram.pdf")
}

}
\seealso{
\itemize{
\item \code{\link[=getDendro]{getDendro()}} to generate dendrograms for \code{plotDendro()}.
}
}
