% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Investigate_Functional_Enrichments.R
\name{plotEnrichment}
\alias{plotEnrichment}
\title{Plot Functional Enrichment Results}
\usage{
plotEnrichment(
  enrichment,
  nTerms = 15,
  fill = "#132B43",
  xlim = NULL,
  nBreaks = 4,
  axis.title.x.size = 20,
  axis.text.x.size = 16,
  axis.text.y.size = 16,
  save = TRUE,
  file = "Module_Enrichment_Plot.pdf",
  width = 8,
  height = 6,
  verbose = TRUE
)
}
\arguments{
\item{enrichment}{A \code{data.frame} of functional enrichment results,
typically obtained from \code{\link[=enrichModule]{enrichModule()}}.}

\item{nTerms}{A \code{numeric(1)} specifying the number of terms to include in
the plot.}

\item{fill}{A \code{character(1)} giving the color of the bars.}

\item{xlim}{A \code{numeric(2)} specifying the limits of the x-axis.}

\item{nBreaks}{A \code{numeric(1)} indicating the number of breaks to use in
the x-axis.}

\item{axis.title.x.size}{A \code{numeric(1)} with the size of the x-axis title.}

\item{axis.text.x.size}{A \code{numeric(1)} giving the size of the x-axis text.}

\item{axis.text.y.size}{A \code{numeric(1)} giving the size of the y-axis text.}

\item{save}{A \code{logical(1)} indicating whether to save the plot.}

\item{file}{A \code{character(1)} giving the file name (.pdf) for the saved
plot.}

\item{width}{A \code{numeric(1)} specifying the width in inches of the saved
plot.}

\item{height}{A \code{numeric(1)} specifying the height in inches of the
saved plot.}

\item{verbose}{A \code{logical(1)} indicating whether messages should be
printed.}
}
\value{
A \code{ggplot} object.
}
\description{
\code{plotEnrichment()} takes a \code{data.frame} of enrichment results from
\code{\link[=enrichModule]{enrichModule()}}, plots the log p-values in a bar plot, and saves it as a .pdf.
}
\details{
\code{plotEnrichment()} is designed to be used in combination with
\code{\link[=enrichModule]{enrichModule()}}. The top 15 gene sets are plotted by default, but this can
be expanded if needed. A \code{ggplot} object is produced and can be edited
outside of this function if desired.
}
\examples{
\dontrun{

# Get Comethylation Modules
modules <- getModules(methAdj, power = sft$powerEstimate, regions = regions,
                      corType = "pearson", file = "Modules.rds")

# Annotate Modules
regionsAnno <- annotateModule(regions, module = "bisque4",
                              genome = "hg38",
                              file = "Annotated_bisque4_Module_Regions.txt")
geneList_bisque4 <- getGeneList(regionsAnno, module = "bisque4")

# Analyze Functional Enrichment
ontologies <- listOntologies("hg38", version = "4.0.4")
enrich_bisque4 <- enrichModule(regions, module = "bisque4",
                               genome = "hg38",
                               file = "bisque4_Module_Enrichment.txt")
plotEnrichment(enrich_bisque4,
               file = "bisque4_Module_Enrichment_Plot.pdf")
}

}
\seealso{
\itemize{
\item \code{\link[=getModules]{getModules()}} to build a comethylation network and identify
modules of comethylated regions.
\item \code{\link[=annotateModule]{annotateModule()}} and \code{\link[=getGeneList]{getGeneList()}} to annotate a set of
regions with genes and regulatory context and then extract
the gene symbols or IDs.
\item \code{\link[=listOntologies]{listOntologies()}} and \code{\link[=enrichModule]{enrichModule()}}, to investigate
functional enrichment of module regions with GREAT.
}
}
