% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Explore_Module_Trait_Correlations.R
\name{plotMEtraitCor}
\alias{plotMEtraitCor}
\title{Plot a Heatmap of Correlations Between Module Eigennodes and Traits}
\usage{
plotMEtraitCor(
  MEtraitCor,
  moduleOrder = 1:length(unique(MEtraitCor$module)),
  traitOrder = 1:length(unique(MEtraitCor$trait)),
  topOnly = FALSE,
  nTop = 15,
  p = 0.05,
  label.type = c("star", "p"),
  label.size = 8,
  label.nudge_y = -0.38,
  colors = blueWhiteRed(100, gamma = 0.9),
  limit = NULL,
  axis.text.size = 12,
  legend.position = c(1.08, 0.915),
  legend.text.size = 12,
  legend.title.size = 16,
  colColorMargins = c(-0.7, 4.21, 1.2, 11.07),
  save = TRUE,
  file = "ME_Trait_Correlation_Heatmap.pdf",
  width = 11,
  height = 9.5,
  verbose = TRUE
)
}
\arguments{
\item{MEtraitCor}{A \code{data.frame} of correlation statistics for
module-trait pairs, typically generated by \code{\link[=getMEtraitCor]{getMEtraitCor()}}.}

\item{moduleOrder}{A \code{numeric} specifying the order of modules in the
heatmap. This must be the same length as the number of unique modules
in \code{MEtraitCor}.}

\item{traitOrder}{A \code{numeric} specifying the order of traits in the
heatmap. This must be the same length as the number of unique traits
in \code{MEtraitCor}.}

\item{topOnly}{A \code{logical(1)} indicating whether to plot only modules
and traits with the most significant correlations, ranked by p-value.}

\item{nTop}{A \code{numeric(1)} specifying the number of top module-trait
associations to include.}

\item{p}{A \code{numeric(1)} giving the threshold for the p-value to assign
a significant correlation.}

\item{label.type}{A \code{character(1)} giving the type of label (star or p)
to use to annotate significant correlations.}

\item{label.size}{A \code{numeric(1)} specifying the size of the label text
for annotating significant correlations.}

\item{label.nudge_y}{A \code{numeric(1)} giving the amount to adjust the
position of the label text on the y-axis.}

\item{colors}{A \code{character} giving a vector of colors to use for the
gradient on the heatmap. The default uses \code{\link[WGCNA:blueWhiteRed]{WGCNA::blueWhiteRed()}}
to generate these colors.}

\item{limit}{A \code{numeric(1)} giving the maximum value (symmetric) for the
heatmap color scale.}

\item{axis.text.size}{A \code{numeric(1)} specifying the size of the text for
the y-axis.}

\item{legend.position}{A \code{numeric(2)} with the position of the legend,
as x-axis, y-axis. May also be a \code{character(1)} indicating
"none", "left", "right", "bottom", or "top".}

\item{legend.text.size}{A \code{numeric(1)} specifying the size of the text
for the legend axis.}

\item{legend.title.size}{A \code{numeric(1)} giving the size of the text for
the legend title.}

\item{colColorMargins}{A \code{numeric(4)} giving the width of the margins
for the column (horizontal) color bar as top, right, bottom, and left.}

\item{save}{A \code{logical(1)} indicating whether to save the plot.}

\item{file}{A \code{character(1)} giving the file name (.pdf) for the saved
plot.}

\item{width}{A \code{numeric(1)} specifying the width in inches of the saved
plot.}

\item{height}{A \code{numeric(1)} specifying the height in inches of the
saved plot.}

\item{verbose}{A \code{logical(1)} indicating whether messages should be
printed.}
}
\value{
A \code{ggplot} object.
}
\description{
\code{plotMEtraitCor()} takes a \code{data.frame} of correlation statistics
for module eigennodes and traits from \code{\link[=getMEtraitCor]{getMEtraitCor()}}, plots it as a
heatmap, and saves it as a .pdf.
}
\details{
\code{plotMEtraitCor()} is designed to be used in combination with
\code{\link[=getMEtraitCor]{getMEtraitCor()}}. Stars or p-values are used to annotate significant
correlations, as defined by the p-value threshold. The heatmap can optionally
be filtered to include only modules and traits with the most significant
associations, ranked by p-value. A \code{ggplot} object is produced and can
be edited outside of this function if desired.
}
\examples{
\dontrun{

# Get Comethylation Modules
modules <- getModules(methAdj, power = sft$powerEstimate, regions = regions,
                      corType = "pearson", file = "Modules.rds")

# Test Correlations between Module Eigennodes and Sample Traits
MEs <- modules$MEs
MEtraitCor <- getMEtraitCor(MEs, colData = colData, corType = "bicor",
                            file = "ME_Trait_Correlation_Stats.txt")

# Examine Correlations between Sample Traits
traitDendro <- getCor(MEs, y = colData, corType = "bicor",
                      robustY = FALSE) \%>\%
        getDendro(transpose = TRUE)
plotDendro(traitDendro, labelSize = 3.5, expandY = c(0.65,0.08),
           file = "Trait_Dendrogram.pdf")

# Visualize Correlations between Module Eigennodes and Sample Traits
moduleDendro <- getDendro(MEs, distance = "bicor")
plotMEtraitCor(MEtraitCor, moduleOrder = moduleDendro$order,
               traitOrder = traitDendro$order,
               file = "ME_Trait_Correlation_Heatmap.pdf")
plotMEtraitCor(MEtraitCor, moduleOrder = moduleDendro$order,
               traitOrder = traitDendro$order, topOnly = TRUE,
               label.type = "p", label.size = 4, label.nudge_y = 0,
               legend.position = c(1.11, 0.795),
               colColorMargins = c(-1,4.75,0.5,10.1),
               file = "Top_ME_Trait_Correlation_Heatmap.pdf", width = 8.5,
               height = 4.25)
}

}
\seealso{
\itemize{
\item \code{\link[=getModules]{getModules()}} to build a comethylation network and identify
modules of comethylated regions.
\item \code{\link[=getMEtraitCor]{getMEtraitCor()}} to calculate ME-trait correlations.
\item \code{\link[=getCor]{getCor()}} to calculate correlation coefficients.
\item \code{\link[=getDendro]{getDendro()}} and \code{\link[=plotDendro]{plotDendro()}} to generate and visualize
dendrograms.
}
}
