% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Explore_Module_Trait_Correlations.R
\name{plotMEtraitScatter}
\alias{plotMEtraitScatter}
\title{Visualize a Module Eigennode - Trait Correlation as a Scatter Plot}
\usage{
plotMEtraitScatter(
  ME,
  trait,
  color = "#132B43",
  xlim = NULL,
  ylim = NULL,
  nBreaks = 4,
  point.size = 2.5,
  axis.title.size = 20,
  axis.text.size = 16,
  xlab = "Trait",
  ylab = "Module Eigennode",
  save = TRUE,
  file = "ME_Trait_Scatterplot.pdf",
  width = 6,
  height = 6,
  verbose = TRUE
)
}
\arguments{
\item{ME}{A \code{numeric} of module eigennode values. \code{ME} must be in the
same order as \code{trait}.}

\item{trait}{A \code{numeric} of continuous sample trait values.}

\item{color}{A \code{character(1)} giving the color of the points.}

\item{xlim}{A \code{numeric(2)} specifying the limits of the x-axis.}

\item{ylim}{A \code{numeric(2)} specifying the limits of the y-axis.}

\item{nBreaks}{A \code{numeric(1)} giving the number of breaks for both axes.}

\item{point.size}{A \code{numeric(1)} indicating the size of the points.}

\item{axis.title.size}{A \code{numeric(1)} indicating the size of the title
text for both axes.}

\item{axis.text.size}{A \code{numeric(1)} specifying the size of the text for
both axes.}

\item{xlab}{A \code{character(1)} giving the x-axis title.}

\item{ylab}{A \code{character(1)} giving the y-axis title.}

\item{save}{A \code{logical(1)} indicating whether to save the plot.}

\item{file}{A \code{character(1)} giving the file name (.pdf) for the saved
plot.}

\item{width}{A \code{numeric(1)} specifying the width in inches of the saved
plot.}

\item{height}{A \code{numeric(1)} specifying the height in inches of the
saved plot.}

\item{verbose}{A \code{logical(1)} indicating whether messages should be
printed.}
}
\value{
A \code{ggplot} object.
}
\description{
\code{plotMEtraitScatter()} takes a \code{vector} of module eigennode values
and a \code{vector} of continuous sample trait values, generates a scatter plot,
and then saves it as a .pdf. \code{ME} and \code{trait} must be in the same
order.
}
\details{
The values in \code{ME} and \code{trait} are plotted as points along with a
smoothed line with a shaded 95\% confidence interval. The smoothed line is fit
using robust regression as implemented by \code{\link[MASS:rlm]{MASS::rlm()}}. A \code{ggplot}
object is produced and can be edited outside of this function if desired.
}
\examples{
\dontrun{

# Get Comethylation Modules
modules <- getModules(methAdj, power = sft$powerEstimate, regions = regions,
                      corType = "pearson", file = "Modules.rds")

# Test Correlations between Module Eigennodes and Sample Traits
MEs <- modules$MEs
MEtraitCor <- getMEtraitCor(MEs, colData = colData, corType = "bicor",
                            file = "ME_Trait_Correlation_Stats.txt")
plotMEtraitCor(MEtraitCor, moduleOrder = moduleDendro$order,
               traitOrder = traitDendro$order,
               file = "ME_Trait_Correlation_Heatmap.pdf")

# Explore Individual ME-Trait Correlations
plotMEtraitDot(MEs$bisque4, trait = colData$Diagnosis_ASD,
               traitCode = c("TD" = 0, "ASD" = 1),
               colors = c("TD" = "#3366CC", "ASD" = "#FF3366"),
               ylim = c(-0.2,0.2), xlab = "Diagnosis",
               ylab = "Bisque 4 Module Eigennode",
               file = "bisque4_ME_Diagnosis_Dotplot.pdf")
plotMEtraitScatter(MEs$paleturquoise, trait = colData$Gran,
                   ylim = c(-0.15,0.15), xlab = "Granulocytes",
                   ylab = "Pale Turquoise Module Eigennode",
                   file = "paleturquoise_ME_Granulocytes_Scatterplot.pdf")
regions <- modules$regions
plotMethTrait("bisque4", regions = regions, meth = meth,
              trait = colData$Diagnosis_ASD,
              traitCode = c("TD" = 0, "ASD" = 1),
              traitColors = c("TD" = "#3366CC", "ASD" = "#FF3366"),
              trait.legend.title = "Diagnosis",
              file = "bisque4_Module_Methylation_Diagnosis_Heatmap.pdf")
}

}
\seealso{
\itemize{
\item \code{\link[=getModules]{getModules()}} to build a comethylation network and identify
modules of comethylated regions.
\item \code{\link[=getMEtraitCor]{getMEtraitCor()}} and \code{\link[=plotMEtraitCor]{plotMEtraitCor()}} to calculate and
visualize all ME-trait correlations.
\item \code{\link[=plotMEtraitDot]{plotMEtraitDot()}} and \code{\link[=plotMethTrait]{plotMethTrait()}} for other methods to
visualize a single ME-trait correlation.
}
}
