% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_Comethylation_Network.R
\name{plotRegionDendro}
\alias{plotRegionDendro}
\title{Plot Region Dendrograms}
\usage{
plotRegionDendro(
  modules,
  save = TRUE,
  file = "Region_Dendrograms.pdf",
  width = 11,
  height = 4.25,
  verbose = TRUE
)
}
\arguments{
\item{modules}{A \code{list} of module assignments and statistics produced by
\code{\link[=getModules]{getModules()}}.}

\item{save}{A \code{logical(1)} indicating whether to save the plot.}

\item{file}{A \code{character(1)} giving the file name (.pdf) for the saved
plot.}

\item{width}{A \code{numeric(1)} specifying the width in inches of the saved
plot.}

\item{height}{A \code{numeric(1)} specifying the height in inches of the
saved plot.}

\item{verbose}{A \code{logical(1)} indicating whether messages should be
printed.}
}
\value{
None, produces a plot as a side effect.
}
\description{
\code{plotRegionDendro()} extracts plotting data from a \code{modules} list,
plots a region dendrogram with module assignments, and then saves it as a
.pdf.
}
\details{
\code{plotRegionDendro()} is designed to be used in combination with
\code{\link[=getModules]{getModules()}}. This function does not produce a \code{ggplot} object, but
instead uses \code{\link[WGCNA:plotDendroAndColors]{WGCNA::plotDendroAndColors()}} to plot the dendrogram.
}
\examples{
\dontrun{

# Get Comethylation Modules
modules <- getModules(methAdj, power = sft$powerEstimate, regions = regions,
                      corType = "pearson", file = "Modules.rds")

# Visualize Comethylation Modules
plotRegionDendro(modules, file = "Region_Dendrograms.pdf")
BED <- getModuleBED(modules$regions, file = "Modules.bed")
}

}
\seealso{
\itemize{
\item \code{\link[=getModules]{getModules()}} to build a comethylation network and identify
modules of comethylated regions.
\item \code{\link[=getModuleBED]{getModuleBED()}} to visualize genomic locations and module
assignments.
\item \code{\link[=getDendro]{getDendro()}} and \code{\link[=plotDendro]{plotDendro()}} to generate and visualize
dendrograms for samples, modules, and traits.
}
}
