% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_and_Filter_Regions.R
\name{plotRegionTotals}
\alias{plotRegionTotals}
\title{Visualize Region Totals at Different Cutoffs}
\usage{
plotRegionTotals(
  regionTotals,
  nBreaks = 4,
  legend.position = c(1.08, 0.897),
  save = TRUE,
  file = "Region_Totals.pdf",
  width = 11,
  height = 11,
  verbose = TRUE
)
}
\arguments{
\item{regionTotals}{A \code{data.frame}, output from \code{\link[=getRegionTotals]{getRegionTotals()}}.}

\item{nBreaks}{A \code{numeric(1)} specifying the number of breaks used for
both axes and the legend.}

\item{legend.position}{A \code{numeric(2)} specifying the position of the
legend, as x-axis, y-axis. May also be a \code{character(1)}
indicating "none", "left", "right", "bottom", or "top".}

\item{save}{A \code{logical(1)} indicating whether to save the plot.}

\item{file}{A \code{character(1)} giving the file name (.pdf) for the saved
plot.}

\item{width}{A \code{numeric(1)} specifying the width in inches of the saved
plot.}

\item{height}{A \code{numeric(1)} specifying the height in inches of the
saved plot.}

\item{verbose}{A \code{logical(1)} indicating whether messages should be
printed.}
}
\value{
A \code{ggplot} object.
}
\description{
\code{plotRegionTotals()} plots the total number of regions, width, and total
number of CpGs remaining after filtering by different combinations of
\code{covMin} and \code{methSD} in a line plot and then saves it as a .pdf.
}
\details{
\code{plotRegionTotals()} is designed to be used in combination with
\code{\link[=getRegionTotals]{getRegionTotals()}}. A \code{ggplot} object is produced and can be edited
outside of this function if desired.
}
\examples{
\dontrun{

# Call Regions
regions <- getRegions(bs, file = "Unfiltered_Regions.txt")
plotRegionStats(regions, maxQuantile = 0.99,
                file = "Unfiltered_Region_Plots.pdf")
plotSDstats(regions, maxQuantile = 0.99,
            file = "Unfiltered_SD_Plots.pdf")

# Examine Region Totals at Different Cutoffs
regionTotals <- getRegionTotals(regions, file = "Region_Totals.txt")
plotRegionTotals(regionTotals, file = "Region_Totals.pdf")

# Filter Regions
regions <- filterRegions(regions, covMin = 10, methSD = 0.05,
                         file = "Filtered_Regions.txt")
plotRegionStats(regions, maxQuantile = 0.99,
                file = "Filtered_Region_Plots.pdf")
plotSDstats(regions, maxQuantile = 0.99,
            file = "Filtered_SD_Plots.pdf")
}

}
\seealso{
\itemize{
\item \code{\link[=getRegions]{getRegions()}} to generate the set of regions.
\item \code{\link[=plotRegionStats]{plotRegionStats()}}, \code{\link[=plotSDstats]{plotSDstats()}}, and \code{\link[=getRegionTotals]{getRegionTotals()}}
for more help visualizing region characteristics and setting
cutoffs for filtering.
\item \code{\link[=filterRegions]{filterRegions()}} for filtering regions by minimum coverage and
methylation standard deviation.
}
}
