% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_and_Filter_Regions.R
\name{plotSDstats}
\alias{plotSDstats}
\title{Plot Heatmaps of Region Standard Deviation vs Features}
\usage{
plotSDstats(
  regions,
  maxQuantile = 1,
  bins = 30,
  nBreaks = 4,
  legend.position = c(1.09, 0.9),
  save = TRUE,
  file = "SD_Plots.pdf",
  width = 8.5,
  height = 8.5,
  verbose = TRUE
)
}
\arguments{
\item{regions}{A \code{data.frame} output from \code{\link[=getRegions]{getRegions()}} giving the set
of regions and statistics for each region.}

\item{maxQuantile}{A \code{numeric(1)} giving the maximum quantile of each
feature to plot.}

\item{bins}{A \code{numeric(1)} specifying the number of bins for both axes
in each heatmap.}

\item{nBreaks}{A \code{numeric(1)} specifying the number of breaks for both
axes.}

\item{legend.position}{A \code{numeric(2)} specifying the position of the
legend, as x-axis, y-axis. May also be a \code{character(1)}
indicating "none", "left", "right", "bottom", or "top".}

\item{save}{A \code{logical(1)} indicating whether to save the plot.}

\item{file}{A \code{character(1)} giving the file name (.pdf) for the plot.}

\item{width}{A \code{numeric(1)} specifying the width in inches of the saved
plot.}

\item{height}{A \code{numeric(1)} specifying the height in inches of the saved
plot.}

\item{verbose}{A \code{logical(1)} indicating whether messages should be
printed.}
}
\value{
A \code{ggplot} object.
}
\description{
\code{plotSDstats()} takes a set of regions from \code{\link[=getRegions]{getRegions()}}, generates
heatmaps of methylation standard deviation against region features, and saves
it as a pdf. Compared features include number of CpGs, minimum coverage, mean
coverage, and mean methylation.
}
\details{
It's recommended examine these plots before and after filtering to ensure
removal of regions with high variability due to insufficient data. Plots are
heatmaps of 2D bin counts, with the color indicating the number of regions in
that bin on the log10 scale. A \code{ggplot} object is produced and can be
edited outside of this function if desired.
}
\examples{
\dontrun{

# Call Regions
regions <- getRegions(bs, file = "Unfiltered_Regions.txt")
plotRegionStats(regions, maxQuantile = 0.99,
                file = "Unfiltered_Region_Plots.pdf")
plotSDstats(regions, maxQuantile = 0.99,
            file = "Unfiltered_SD_Plots.pdf")

# Examine Region Totals at Different Cutoffs
regionTotals <- getRegionTotals(regions, file = "Region_Totals.txt")
plotRegionTotals(regionTotals, file = "Region_Totals.pdf")

# Filter Regions
regions <- filterRegions(regions, covMin = 10, methSD = 0.05,
                         file = "Filtered_Regions.txt")
plotRegionStats(regions, maxQuantile = 0.99,
                file = "Filtered_Region_Plots.pdf")
plotSDstats(regions, maxQuantile = 0.99,
            file = "Filtered_SD_Plots.pdf")
}

}
\seealso{
\itemize{
\item \code{\link[=getRegions]{getRegions()}} to generate the set of regions.
\item \code{\link[=plotRegionStats]{plotRegionStats()}}, \code{\link[=getRegionTotals]{getRegionTotals()}}, and \code{\link[=plotRegionTotals]{plotRegionTotals()}}
for more help visualizing region characteristics and setting
cutoffs for filtering.
\item \code{\link[=filterRegions]{filterRegions()}} for filtering regions by minimum coverage and
methylation standard deviation.
}
}
