% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build_Comethylation_Network.R
\name{plotSoftPower}
\alias{plotSoftPower}
\title{Plot Soft Power Fit and Connectivity}
\usage{
plotSoftPower(
  sft,
  pointCol = "#132B43",
  lineCol = "red",
  nBreaks = 4,
  save = TRUE,
  file = "Soft_Power_Plots.pdf",
  width = 8.5,
  height = 4.25,
  verbose = TRUE
)
}
\arguments{
\item{sft}{A \code{list} produced by \code{\link{getSoftPower()}} with two
elements: \code{powerEstimate} and \code{fitIndices}.}

\item{pointCol}{A \code{character(1)} specifying the color of the points.}

\item{lineCol}{A \code{character(1)} giving the color of line and label for
the estimated soft power threshold for scale-free topology.}

\item{nBreaks}{A \code{numeric(1)} specifying the number of breaks used for
both axes.}

\item{save}{A \code{logical(1)} indicating whether to save the plot.}

\item{file}{A \code{character(1)} giving the file name (.pdf) for the saved
plot.}

\item{width}{A \code{numeric(1)} specifying the width in inches of the saved
plot.}

\item{height}{A \code{numeric(1)} specifying the height in inches of the
saved plot.}

\item{verbose}{A \code{logical(1)} indicating whether messages should be
printed.}
}
\value{
A \code{ggplot} object.
}
\description{
\code{plotSoftPower()} visualizes scale-free topology fit and mean connectivity
for multiple soft power thresholds as a scatterplot, and then saves it as a
.pdf.
}
\details{
\code{plotSoftPower()} is designed to be used in combination with
\code{\link{getSoftPower()}}. A \code{ggplot} object is produced and can be
edited outside of this function if desired.
}
\examples{
\dontrun{

# Get Methylation Data
meth <- getRegionMeth(regions, bs = bs, file = "Region_Methylation.rds")

# Adjust Methylation Data for PCs
mod <- model.matrix(~1, data = pData(bs))
methAdj <- adjustRegionMeth(meth, mod = mod,
                            file = "Adjusted_Region_Methylation.rds")

# Select Soft Power Threshold
sft <- getSoftPower(methAdj, corType = "pearson", file = "Soft_Power.rds")
plotSoftPower(sft, file = "Soft_Power_Plots.pdf")

# Get Comethylation Modules
modules <- getModules(methAdj, power = sft$powerEstimate, regions = regions,
                      corType = "pearson", file = "Modules.rds")
}

}
\seealso{
\itemize{
\item \code{\link{getRegionMeth()}} and \code{\link{adjustRegionMeth()}}
to extract methylation data and then adjust it for the top
principal components.
\item \code{\link{getSoftPower()}} to calculate the best soft-thresholding
power and fit indices for scale-free topology.
\item \code{\link{getModules()}} to build a comethylation network and
identify modules of comethylated regions.
}
}
