% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_and_Filter_CpGs.R
\name{filterCpGs}
\alias{filterCpGs}
\title{Filter BSseq Objects by Coverage}
\usage{
filterCpGs(
  bs,
  cov = 2,
  perSample = 0.75,
  save = TRUE,
  file = "Filtered_BSseq.rds",
  verbose = TRUE
)
}
\arguments{
\item{bs}{A \code{\link[bsseq]{BSseq}} object.}

\item{cov}{A \code{numeric(1)} specifying the minimum number of reads
overlapping a CpG for it to be included in the total.}

\item{perSample}{A \code{numeric(1)} specifying the minimum percent of
samples with \code{cov} reads at a CpG for it to be included in the
total.}

\item{save}{A \code{logical(1)} indicating whether to save the BSseq object.}

\item{file}{A \code{character(1)} giving the file name (.rds) for the saved
BSseq object.}

\item{verbose}{A \code{logical(1)} indicating whether messages should be
printed.}
}
\value{
A \code{\link[bsseq]{BSseq}} object.
}
\description{
\code{filterCpGs()} subsets a \code{\link[bsseq]{BSseq}} object to include
only those CpGs meeting \code{cov} and \code{perSample} cutoffs and
then saves it as a .rds file.
}
\details{
\code{filterCpGs()} is designed to be used after \code{cov} and
\code{perSample} arguments have been optimized by
\code{\link{getCpGtotals()}} and \code{\link{plotCpGtotals()}}.
}
\examples{
\dontrun{

# Read Bismark CpG Reports
colData <- read.xlsx("sample_info.xlsx", rowNames = TRUE)
bs <- getCpGs(colData, file = "Unfiltered_BSseq.rds")

# Examine CpG Totals at Different Cutoffs
CpGtotals <- getCpGtotals(bs, file = "CpG_Totals.txt")
plotCpGtotals(CpGtotals, file = "CpG_Totals.pdf")

# Filter BSobject
bs <- filterCpGs(bs, cov = 2, perSample = 0.75, file = "Filtered_BSseq.rds")
}

}
\seealso{
\itemize{
\item \code{\link{getCpGtotals()}} and \code{\link{plotCpGtotals()}}
for help with deciding coverage cutoffs.
\item \code{\link{filterCpGs()}} to filter the
\code{\link[bsseq]{BSseq}} object.
\item \code{\link{getCpGs()}} to generate the
\code{\link[bsseq]{BSseq}} object from individual Bismark CpG
reports.
}
}
